apps/workers/requirements.txt
# Core framework
fastapi==0.104.1
uvicorn[standard]==0.24.0
pydantic==2.4.2
pydantic-settings==2.0.3

# Database
asyncpg==0.29.0
sqlalchemy[asyncio]==2.0.23
alembic==1.12.1
pgvector==0.2.4

# Message queue & caching
redis[hiredis]==5.0.1
nats-py==2.6.0

# AI & ML
langchain==0.0.335
langchain-openai==0.0.2
langchain-anthropic==0.0.1
openai==1.3.5
anthropic==0.7.7
sentence-transformers==2.2.2
transformers==4.35.2
torch==2.1.1
numpy==1.25.2
scikit-learn==1.3.2

# TTS & Audio
elevenlabs==0.2.26
azure-cognitiveservices-speech==1.34.0
pydub==0.25.1
librosa==0.10.1

# Document processing
pypdf2==3.0.1
python-docx==1.1.0
python-pptx==0.6.23
pillow==10.1.0
opencv-python==4.8.1.78

# OCR & Vision
pytesseract==0.3.10
easyocr==1.7.0

# 3D & Geometry
trimesh==4.0.5
open3d==0.18.0

# HTTP & Networking
httpx==0.25.2
aiofiles==23.2.1

# Observability
opentelemetry-api==1.21.0
opentelemetry-sdk==1.21.0
opentelemetry-exporter-otlp==1.21.0
opentelemetry-instrumentation-fastapi==0.42b0
opentelemetry-instrumentation-sqlalchemy==0.42b0
opentelemetry-instrumentation-redis==0.42b0
prometheus-client==0.19.0
structlog==23.2.0

# Security & Validation
cryptography==41.0.7
python-jose[cryptography]==3.3.0
passlib[bcrypt]==1.7.4

# Development
pytest==7.4.3
pytest-asyncio==0.21.1
pytest-cov==4.1.0
black==23.11.0
isort==5.12.0
flake8==6.1.0
mypy==1.7.1
